pyproject.toml
[build-system]
requires = [ "poetry-core",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "backend"
version = "2025.1.2"
description = ""
authors = [ "Prasanna S R <prasanna@beehyv.com>",]
readme = "README.md"
package-mode = false

[tool.black]
line-length = 120

[tool.isort]
skip = [ "__init__.py",]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.113.0"
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
gunicorn = "^23.0.0"
psycopg-binary = "^3.2.1"
pydantic-settings = "^2.4.0"
asyncpg = "^0.29.0"
python-json-logger = "^2.0.7"
bcrypt = "^4.2.0"
jwt = "^1.3.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
celery = "^5.4.0"
flower = "^2.0.1"
httpx = "^0.27.2"
gevent = "^24.2.1"
contextvars = "^2.4"
email-validator = "^2.2.0"
python-multipart = "^0.0.9"
pypdf2 = "^3.0.1"
langchain-core = "^0.3.6"
langchain-postgres = "^0.0.12"
jinja2 = "^3.1.4"
mailjet-rest = "^1.3.4"
asgiref = "^3.8.1"
celery-pool-asyncio = "^0.2.0"
python-dotenv = "^1.0.1"
opentelemetry-instrumentation-requests = "^0.48b0"
opentelemetry-exporter-prometheus = "^0.48b0"
opentelemetry-instrumentation-fastapi = "^0.48b0"
opentelemetry-instrumentation-sqlalchemy = "^0.48b0"
opentelemetry-instrumentation = "^0.48b0"
opentelemetry-instrumentation-threading = "^0.48b0"
opentelemetry-exporter-jaeger = "^1.21.0"
opentelemetry-exporter-otlp = "^1.27.0"
psycopg2-binary = "^2.9.9"
langchain-community = "^0.3.1"
langchain-openai = "^0.2.1"
langchain-google-vertexai = "^2.0.1"
google-cloud-aiplatform = "^1.68.0"
sse-starlette = "^2.1.3"
websockets = "^13.1"
pydub = "^0.25.1"
ffmpeg = "^1.4"
lxml = "^5.3.0"
beautifulsoup4 = "^4.12.3"
html-chunking = "^0.0.4"
clickhouse-connect = "^0.8.10"
clickhouse-driver = "^0.2.9"

[tool.poetry.dependencies.uvicorn]
extras = [ "standard",]
version = "^0.30.6"

[tool.poetry.dependencies.psycopg]
extras = [ "binaries",]
version = "^3.2.1"

[tool.poetry.dependencies.pydantic]
extras = [ "email",]
version = "2.9.2"

[tool.poetry.group.dev.dependencies]
devtools = "^0.12.2"
safety = "^3.2.7"
pyupgrade = "^3.17.0"
ruff = "^0.6.3"
tryceratops = "^2.4.0"
locust = "^2.31.5"
pytest = "^8.3.2"
flake8 = "^7.1.1"
mypy = "^1.11.2"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"
watchdog = "^5.0.3"
twilio = "^9.3.7"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
pre-commit = "^4.0.1"
black = "^24.10.0"
